Use MessageStream in the ingestor
