+++ catalogue_pipeline/ingestor/src/test/scala/uk/ac/wellcome/platform/ingestor/services/IngestorWorkerServiceTest.scala
@@ -186,14 +186,13 @@ class IngestorWorkerServiceTest
     testWith: TestWith[IngestorWorkerService, R]): R = {
     withLocalSqsQueue { queue =>
       withLocalS3Bucket { bucket =>
-        withMessageReader[IdentifiedWork, R](bucket, queue) { messageReader =>
+        withMessageStream[IdentifiedWork, R](actorSystem, bucket, queue, metricsSender) { messageStream =>
           val service = new IngestorWorkerService(
             esIndexV1 = esIndexV1,
             esIndexV2 = esIndexV2,
             identifiedWorkIndexer = workIndexer,
-            messageReader = messageReader,
-            system = actorSystem,
-            metrics = metricsSender
+            messageStream = messageStream,
+            system = actorSystem
           )
 
           testWith(service)
